Replace MySql with PostgreSql

+++ car-service/src/main/resources/schema.sql
@@ -9,6 +9,6 @@ CREATE TABLE car
 (
     code VARCHAR(255) NOT NULL PRIMARY KEY,
     model VARCHAR(255) NOT NULL,
-    rent_per_kilo DOUBLE NOT NULL,
+    rent_per_kilo DOUBLE PRECISION NOT NULL,
     manufacturer_name VARCHAR(255) NOT NULL REFERENCES manufacturer(name)
 );
